tools/aws_secretsmanager_rotation/requirements.txt
pyyaml>=6.0
boto3
pyjwt
cdk-proxy-api-client>=0.5.1
compose-x-common>=1.3
